Time: 4 ms (97.51%) | Memory: 44.9 MB (54.66%) - LeetSync

+++ 155-min-stack/min-stack.java
@@ -4,7 +4,7 @@ class MinStack {
     long min=0;
     public void push(int val) {
         if(s.isEmpty()){
-            min=val;
+            min=(long)val;
             s.push((long)val);
         }
         else{
